+++ spotifywebapipython/spotifyauthtoken.py
@@ -133,6 +133,25 @@ class SpotifyAuthToken:
     
 
     @property
+    def HeaderKey(self) -> str:
+        """
+        Returns a string containing the header key to assign the authorization token value to.
+        This will always return 'Authorization'.
+        """
+        return 'Authorization'
+
+
+    @property
+    def HeaderValue(self) -> str:
+        """
+        Returns the value portion of the authorization header, in the form of '{tokentype} {token}'.
+
+        Example: `Bearer BQCylDIef4JiH ...'
+        """
+        return '{tokentype} {token}'.format(tokentype=self.TokenType, token=self.AccessToken)
+
+
+    @property
     def IsExpired(self) -> bool:
         """ 
         Returns true if the token has expired; otherwise, False if not expired.
@@ -201,17 +220,6 @@ class SpotifyAuthToken:
         """
         return self._TokenType
     
-
-    def GetHeaders(self) -> dict:
-        """
-        Returns a dictionary containing an 'Authorization': 'Bearer {token}' value,
-        which can be used when makeing requests to the Spotify Web API.
-        """
-        headers:dict = {
-            'Authorization': '{tokentype} {token}'.format(tokentype=self.TokenType, token=self.AccessToken)
-        }
-        return headers
-
 
     def ToString(self) -> str:
         """